+++ lib/src/openai.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
 import 'dart:io';
-
-import 'package:chat_gpt_sdk/chat_gpt_sdk.dart';
 import 'package:chat_gpt_sdk/src/audio.dart';
 import 'package:chat_gpt_sdk/src/client/client.dart';
 import 'package:chat_gpt_sdk/src/embedding.dart';
@@ -12,17 +10,15 @@ import 'package:chat_gpt_sdk/src/model/chat_complete/response/ChatCTResponse.dar
 import 'package:chat_gpt_sdk/src/model/client/http_setup.dart';
 import 'package:chat_gpt_sdk/src/model/complete_text/request/complete_text.dart';
 import 'package:chat_gpt_sdk/src/model/complete_text/response/complete_response.dart';
-import 'package:chat_gpt_sdk/src/model/edits/request/edit_request.dart';
-import 'package:chat_gpt_sdk/src/model/edits/response/edit_response.dart';
 import 'package:chat_gpt_sdk/src/model/gen_image/request/generate_image.dart';
-import 'package:chat_gpt_sdk/src/model/gen_image/request/variation.dart';
 import 'package:chat_gpt_sdk/src/model/gen_image/response/GenImgResponse.dart';
 import 'package:chat_gpt_sdk/src/model/openai_engine/engine_model.dart';
 import 'package:chat_gpt_sdk/src/model/openai_model/openai_models.dart';
+import 'package:chat_gpt_sdk/src/moderations.dart';
 import 'package:chat_gpt_sdk/src/utils/constants.dart';
+import 'package:chat_gpt_sdk/src/utils/keep_token.dart';
 import 'package:dio/dio.dart';
 import 'package:dio/io.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'client/exception/openai_exception.dart';
 import 'client/interceptor/interceptor_wrapper.dart';
 import 'edit.dart';
@@ -38,12 +34,11 @@ abstract class IOpenAI {
   });
   Stream<ChatCTResponse> onChatCompletionSSE(
       {required ChatCompleteText request});
-  Stream<GenImgResponse> generateImageStream(GenerateImage request);
   Future<GenImgResponse?> generateImage(GenerateImage request);
   void cancelAIGenerate();
 }
 
-const msgDeprecate = "not support in version 2.0.6";
+//const msgDeprecate = "not support in version 2.0.6";
 
 class OpenAI implements IOpenAI {
   OpenAI._();
@@ -57,54 +52,43 @@ class OpenAI implements IOpenAI {
   /// use for access for chat gpt [_token]
   static String? _token;
 
-  static SharedPreferences? _prefs;
-
   ///cancel request
   final _cancel = CancelToken();
-
-  ///new instance prefs for keep my data
-  void _buildShared() async {
-    _prefs = await SharedPreferences.getInstance();
-  }
 
   /// set new token
   void setToken(String token) async {
     _token = token;
-    await _prefs?.setString(kTokenKey, token);
+    TokenBuilder.build.setToken(token);
   }
 
-  String getToken() => "$_token";
+  String get token => "${TokenBuilder.build.token}";
 
   ///build environment for openai [build]
   ///setup http client
   ///setup logger\
   @override
   OpenAI build({String? token, HttpSetup? baseOption, bool isLog = false}) {
-    _buildShared();
-    Future.delayed(Duration(seconds: 1), () {
-      if ("$token".isEmpty) throw MissionTokenException();
-      final setup = baseOption == null ? HttpSetup() : baseOption;
-      setToken(token!);
+    if ("$token".isEmpty) throw MissionTokenException();
+    final setup = baseOption == null ? HttpSetup() : baseOption;
+    setToken(token!);
 
-      final dio = Dio(BaseOptions(
-          sendTimeout: setup.sendTimeout,
-          connectTimeout: setup.connectTimeout,
-          receiveTimeout: setup.receiveTimeout));
-      if (setup.proxy.isNotEmpty) {
-        dio.httpClientAdapter = IOHttpClientAdapter()
-          ..onHttpClientCreate = (client) {
-            client.findProxy = (uri) {
-              /// "PROXY localhost:7890"
-              return setup.proxy;
-            };
-            return client;
+    final dio = Dio(BaseOptions(
+        sendTimeout: setup.sendTimeout,
+        connectTimeout: setup.connectTimeout,
+        receiveTimeout: setup.receiveTimeout));
+    if (setup.proxy.isNotEmpty) {
+      dio.httpClientAdapter = IOHttpClientAdapter()
+        ..onHttpClientCreate = (client) {
+          client.findProxy = (uri) {
+            /// "PROXY localhost:7890"
+            return setup.proxy;
           };
-      }
-      dio.interceptors.add(InterceptorWrapper(_prefs, token));
+          return client;
+        };
+    }
+    dio.interceptors.add(InterceptorWrapper());
 
-      _client = OpenAIClient(dio: dio, isLogging: isLog);
-      setToken(token);
-    });
+    _client = OpenAIClient(dio: dio, isLogging: isLog);
     return instance;
   }
 
@@ -143,59 +127,9 @@ class OpenAI implements IOpenAI {
     });
   }
 
-  ///### About Method [onCompleteText]
-  /// - Answer questions based on existing knowledge.
-  /// - Create code to call the Stripe API using natural language.
-  /// - Classify items into categories via example.
-  /// - look more
-  /// https://beta.openai.com/examples
-  @Deprecated(msgDeprecate)
-  /**
-   * try it
-   * ```dart
-   *   final request = CompleteText(
-      prompt: translateEngToThai(word: _txtWord.text.toString()),
-      maxTokens: 200,
-      model: Model.TextDavinci3);
 
-      final response = openAI.onCompletion(request: request).asStream();
-   * ```
-   */
-  Stream<CTResponse?> onCompletionStream({required CompleteText request}) {
-    _completeText(request: request);
-    return _completeControl!.stream;
-  }
-
-  StreamController<CTResponse>? _completeControl =
-      StreamController<CTResponse>.broadcast();
-  void _completeText({required CompleteText request}) {
-    _client
-        .postStream("$kURL$kCompletion", _cancel, request.toJson())
-        .listen((rawData) {
-      if (rawData.statusCode != HttpStatus.ok) {
-        _client.log.log("code: ${rawData.statusCode}, error: ${rawData.data}");
-        _completeControl
-          ?..sink
-          ..addError(
-              "complete error: ${rawData.statusMessage} code: ${rawData.statusCode} data: ${rawData.data}");
-      } else {
-        _client.log.log("============= success ==================");
-        _completeControl
-          ?..sink
-          ..add(CTResponse.fromJson(rawData.data));
-      }
-    }).onError((err) {
-      if (err is DioError) {
-        _client.log.error(err, err.stackTrace);
-        _completeControl
-          ?..sink
-          ..addError(
-              "complete error: message :${err.message}\n error body :${err.response?.data}, code: ${err.response?.statusCode}");
-      }
-    });
-  }
-
-  ///Given a chat conversation, the model will return a chat completion response.
+  ///Given a chat conversation,
+  /// the model will return a chat completion response.[onChatCompletion]
   @override
   Future<ChatCTResponse?> onChatCompletion(
       {required ChatCompleteText request}) {
@@ -205,92 +139,6 @@ class OpenAI implements IOpenAI {
     });
   }
 
-  ///Given a chat conversation, the model will return a chat completion response.
-  @Deprecated(msgDeprecate)
-  Stream<ChatCTResponse?> onChatCompletionStream(
-      {required ChatCompleteText request}) {
-    _chatCompleteText(request: request);
-    return _chatCompleteControl!.stream;
-  }
-
-  StreamController<ChatCTResponse>? _chatCompleteControl =
-      StreamController<ChatCTResponse>.broadcast();
-  void _chatCompleteText({required ChatCompleteText request}) {
-    _client
-        .postStream("$kURL$kChatGptTurbo", _cancel, request.toJson())
-        .listen((rawData) {
-      if (rawData.statusCode != HttpStatus.ok) {
-        _client.log.log("code: ${rawData.statusCode}, error: ${rawData.data}");
-        _chatCompleteControl
-          ?..sink
-          ..addError(
-              "chat complete error: ${rawData.statusMessage} code: ${rawData.statusCode} data: ${rawData.data}");
-      } else {
-        _client.log.log("============= success ==================");
-        _chatCompleteControl
-          ?..sink
-          ..add(ChatCTResponse.fromJson(rawData.data));
-      }
-    }).onError((err) {
-      if (err is DioError) {
-        _client.log.error(err, err.stackTrace);
-        _chatCompleteControl
-          ?..sink
-          ..addError(
-              "chat complete error message :${err.message}\n error body :${err.response?.data}, code: ${err.response?.statusCode}");
-      }
-    });
-  }
-
-  ///### close complete stream
-  ///free memory [close]
-  @Deprecated(msgDeprecate)
-  void close() {
-    _completeControl?.close();
-    _chatCompleteControl?.close();
-  }
-
-  ///generate image with prompt [generateImageStream]
-  @override
-  Stream<GenImgResponse> generateImageStream(GenerateImage request) {
-    _generateImage(request);
-    return _genImgController.stream;
-  }
-
-  final _genImgController = StreamController<GenImgResponse>.broadcast();
-  void _generateImage(GenerateImage request) {
-    _client
-        .postStream("$kURL$kGenerateImage", _cancel, request.toJson())
-        .listen((rawData) {
-      if (rawData.statusCode != HttpStatus.ok) {
-        _client.log.log("code: ${rawData.statusCode}, error: ${rawData.data}");
-        _genImgController
-          ..sink
-          ..addError(
-              "generate image error: ${rawData.statusMessage} code: ${rawData.statusCode} data: ${rawData.data}");
-      } else {
-        _client.log.log(
-            "============= success ==================\nresponse data :${rawData.data}");
-        _genImgController
-          ..sink
-          ..add(GenImgResponse.fromJson(rawData.data));
-      }
-    }).onError((err) {
-      if (err is DioError) {
-        _client.log.error(err, err.stackTrace);
-        _genImgController
-          ..sink
-          ..addError(
-              "generate image error: message :${err.message}\nerror body :${err.response?.data}, code: ${err.response?.statusCode}");
-      }
-    });
-  }
-
-  /// close generate image stream[genImgClose]
-  /// free memory
-  void genImgClose() {
-    _genImgController.close();
-  }
 
   ///generate image with prompt
   @override
@@ -301,6 +149,9 @@ class OpenAI implements IOpenAI {
     });
   }
 
+  ///## Support Server Sent Event
+  ///Given a chat conversation,
+  /// the model will return a chat completion response. [onChatCompletionSSE]
   @override
   Stream<ChatCTResponse> onChatCompletionSSE(
       {required ChatCompleteText request}) {
@@ -310,6 +161,12 @@ class OpenAI implements IOpenAI {
     });
   }
 
+  ///## Support Server Sent Event
+  /// - Answer questions based on existing knowledge.
+  /// - Create code to call the Stripe API using natural language.
+  /// - Classify items into categories via example.
+  /// - look more
+  /// https://beta.openai.com/examples .[onChatCompletion]
   @override
   Stream<CTResponse> onCompletionSSE({required CompleteText request}) {
     return _client.sse('$kURL$kCompletion', _cancel,
@@ -334,8 +191,11 @@ class OpenAI implements IOpenAI {
   Audio get audio => Audio(_client);
 
   ///files
- OpenAIFile get file => OpenAIFile(_client);
+  OpenAIFile get file => OpenAIFile(_client);
 
- ///fine-tune
- FineTune get fineTune => FineTune(_client);
+  ///fine-tune
+  FineTune get fineTune => FineTune(_client);
+
+  ///moderations
+  Moderation get moderation => Moderation(_client);
 }